Merge the latest update of admin and super admin java code to latest dev

+++ app/src/main/java/com/example/cpe408finalproject/superAdminDashboard.java
@@ -32,6 +32,7 @@ import com.google.firebase.firestore.QuerySnapshot;
 
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Objects;
 
 public class superAdminDashboard extends AppCompatActivity {
 
@@ -90,6 +91,10 @@ public class superAdminDashboard extends AppCompatActivity {
             }
         });
 
+        Intent intent = getIntent();
+        String username = intent.getStringExtra("username");
+        nameGreeting.setText(username);
+
         addSubmit.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
@@ -125,6 +130,9 @@ public class superAdminDashboard extends AppCompatActivity {
                                                 .add(users);
                                         Log.w(TAG, "User added");
                                         showDialog("User has been added successfully");
+                                        addUsername.setText("");
+                                        addPassword.setText("");
+                                        userTypeRad.clearCheck();
                                     } else {
                                         showDialog("Username has already been taken");
                                     }
@@ -134,7 +142,6 @@ public class superAdminDashboard extends AppCompatActivity {
             }
         }
     });
-
         updateSubmit.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
@@ -153,49 +160,65 @@ public class superAdminDashboard extends AppCompatActivity {
                                 @Override
                                 public void onComplete(@NonNull Task<QuerySnapshot> task) {
                                     if (task.isSuccessful()) {
+                                        DocumentSnapshot document = task.getResult().getDocuments().get(0);
+                                        DocumentReference documentReference = document.getReference();
                                         if(task.getResult().isEmpty()){
                                             Log.w(TAG, "Updating Failed");
                                             showDialog("User does not exist");
                                         }
                                         else
                                         {
-                                            DocumentSnapshot document = task.getResult().getDocuments().get(0);
-                                            DocumentReference documentReference = document.getReference();
-                                            Map<String,Object> users = new HashMap<>();
-                                            users.put("username", newUser);
-                                            users.put("password", password);
-                                            documentReference.update(users)
-                                                    .addOnSuccessListener(new OnSuccessListener<Void>() {
+                                            db.collection("users")
+                                                    .whereEqualTo("username", newUser)
+                                                    .get()
+                                                    .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
                                                         @Override
-                                                        public void onSuccess(Void aVoid) {
-                                                            Log.w(TAG, "Updated user");
-                                                            showDialog("Success! User information updated.");
+                                                        public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                                                            if (task.isSuccessful()){
+                                                                if (task.getResult().isEmpty()){
+                                                                    Map<String,Object> users = new HashMap<>();
+                                                                    users.put("username", newUser);
+                                                                    users.put("password", password);
+                                                                    documentReference.update(users)
+                                                                            .addOnSuccessListener(new OnSuccessListener<Void>() {
+                                                                                @Override
+                                                                                public void onSuccess(Void unused) {
+                                                                                    Log.w(TAG, "Updated user");
+                                                                                    showDialog("Success! User information updated.");
+                                                                                    updateCurrentUsername.setText("");
+                                                                                    updateNewUsername.setText("");
+                                                                                    updateNewPass.setText("");
+                                                                                }
+                                                                            })
+                                                                            .addOnFailureListener(new OnFailureListener() {
+                                                                                @Override
+                                                                                public void onFailure(@NonNull Exception e) {
+                                                                                    Log.w(TAG, "Updated Failed");
+                                                                                    showDialog("Failed! User information not updated.");
+                                                                                }
+                                                                            });
+                                                                } else{ showDialog("Username Already Taken");}
+                                                            }
                                                         }
-                                                    })
-                                                    .addOnFailureListener(new OnFailureListener() {
-                                                        @Override
-                                                        public void onFailure(@NonNull Exception e) {
-                                                            Log.w(TAG, "Updated Failed");
-                                                            showDialog("Failed! User information not updated.");
-                                                        }
-                                                    });}
-                                    } else { showDialog("User Information Error");}
-                                }
+                                                    });
+                                        }
+                                    } else { showDialog("User Information Error");}}
                             });
-
                 }
             }
         });
-
         delSubmit.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
                 String username = delUser.getText().toString();
+                Intent intent = getIntent();
+                String serverUser = intent.getStringExtra("username");
 
                 if (username.isEmpty()) {
                     showDialog("Username cannot be blank!");
-                }
-                else {
+                } else if (username.equals(serverUser)) {
+                    showDialog("Cannot delete own account!");
+                } else {
                     db.collection("users")
                             .whereEqualTo("username", username)
                             .get()
@@ -209,22 +232,30 @@ public class superAdminDashboard extends AppCompatActivity {
                                         }
                                         else {
                                             DocumentSnapshot document = task.getResult().getDocuments().get(0);
-                                            DocumentReference documentReference = document.getReference();
-                                            documentReference.delete()
-                                                    .addOnSuccessListener(new OnSuccessListener<Void>() {
-                                                        @Override
-                                                        public void onSuccess(Void unused) {
-                                                            Log.w(TAG, "Delete success");
-                                                            showDialog("Success! User " + username+ " successfully deleted.");
-                                                        }
-                                                    })
-                                                    .addOnFailureListener(new OnFailureListener() {
-                                                        @Override
-                                                        public void onFailure(@NonNull Exception e) {
-                                                            Log.w(TAG, "Delete Failed");
-                                                            showDialog("Deletion unsuccessful");
-                                                        }
-                                                    });
+                                            String superAdmin = "superAdmin";
+                                            String role = Objects.requireNonNull(document.get("role")).toString();
+                                            if (superAdmin.equals(role)) {
+                                                Log.w(TAG, "User superAdmin");
+                                                showDialog("Cannot delete superAdmin");
+                                            } else {
+                                                DocumentReference documentReference = document.getReference();
+                                                documentReference.delete()
+                                                        .addOnSuccessListener(new OnSuccessListener<Void>() {
+                                                            @Override
+                                                            public void onSuccess(Void unused) {
+                                                                Log.w(TAG, "Delete success");
+                                                                showDialog("Success! User " + username + " successfully deleted.");
+                                                                delUser.setText("");
+                                                            }
+                                                        })
+                                                        .addOnFailureListener(new OnFailureListener() {
+                                                            @Override
+                                                            public void onFailure(@NonNull Exception e) {
+                                                                Log.w(TAG, "Delete Failed");
+                                                                showDialog("Deletion unsuccessful");
+                                                            }
+                                                        });
+                                            }
                                         }
                                     } else { Log.w(TAG, "Unknown Error"); showDialog("Error Detected");}
                                 }
